server/requirements.txt
Flask
gunicorn
numpy
Pillow
tensorflow
